pyproject.toml
[project]
name = "appgarden"
version = "0.1.0"
description = "Deploy web applications to remote servers."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.15.0",
    "pyinfra>=3.0",
    "jinja2>=3.1",
    "rich>=13.0",
    "tomli>=2.0",
    "tomli-w>=1.0",
]

[project.scripts]
appgarden = "appgarden.cli:app_main"

[dependency-groups]
dev = [
    "pytest>=9.0",
    "nblite>=1.1.5",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
ignore-vcs = true
sources = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120
extend-exclude = [
    "__*.pct.py",
    "*.ipynb",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "E402",   # module level import not at top
]

[tool.ruff.lint.per-file-ignores]
"*.pct.py" = [
    "I001", # Unsorted imports
    "E701", # Multiple statements on one line
    "E702", # Multiple statements on one line semicolon
    "E703", # useless-semicolon
]

[tool.ruff.lint.isort]
known-first-party = ["appgarden"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["src/tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short -m 'not integration'"
pythonpath = ["src"]
markers = [
    "integration: integration tests requiring a remote server",
]
